pyproject.toml
[tool.poetry]
name = "pythonproject12"
version = "0.1.0"
description = ""
authors = ["gulnaramari <gulnaramari@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.11.2"
isort = "^5.13.2"
black = "^24.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
)
'''


[tool.isort]
line_length = 119
